Move read images part; Change JPanel into BoardGraphics
